refactor: 브릿지 패턴 적용  #68
- service 추상화와 구현 분리
