+++ competition-service/src/main/java/com/_42195km/msa/competitionservice/application/service/ParticipantService.java
@@ -4,117 +4,21 @@ import java.util.UUID;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import com._42195km.msa.common.exception.CustomBusinessException;
 import com._42195km.msa.competitionservice.application.dto.response.ParticipantAppResponseDto;
 import com._42195km.msa.competitionservice.application.dto.response.SearchParticipantAppResponseDto;
-import com._42195km.msa.competitionservice.application.exception.CompetitionServiceCode;
-import com._42195km.msa.competitionservice.application.mapper.ParticipantMapper;
-import com._42195km.msa.competitionservice.domain.model.Competition;
-import com._42195km.msa.competitionservice.domain.model.CompetitionParticipantMapping;
-import com._42195km.msa.competitionservice.infrastructure.persistence.CompetitionParticipantMappingRepositoryImpl;
-import com._42195km.msa.competitionservice.infrastructure.persistence.CompetitionRepositoryImpl;
-import com._42195km.msa.competitionservice.infrastructure.persistence.ParticipantRepositoryImpl;
 import com._42195km.msa.competitionservice.presentation.dto.request.CancelParticipantRequestDto;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+public interface ParticipantService {
 
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class ParticipantService {
+	Page<ParticipantAppResponseDto> getParticipants(Pageable pageable, UUID competitionId);
 
-	private final ParticipantRepositoryImpl participantRepository;
-	private final CompetitionRepositoryImpl competitionRepository;
-	private final CompetitionParticipantMappingRepositoryImpl mappingRepository;
-	private final ParticipantMapper participantMapper;
+	Page<SearchParticipantAppResponseDto> searchParticipants(String keyword, String searchType, Pageable pageable);
 
-	public Page<ParticipantAppResponseDto> getParticipants(Pageable pageable, UUID competitionId) {
-		try {
-			Competition competition = competitionRepository.findById(competitionId);
-			Page<CompetitionParticipantMapping> participants = mappingRepository.findParticipants(competitionId,
-				pageable);
-			return participantMapper.toParticipantAppResponseDtoPage(participants);
-		} catch (Exception e) {
-			log.error("대회 참가자 검색 실패 로그 : {}", e.getMessage());
-			throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_GET_FAIL);
-		}
-	}
+	Page<SearchParticipantAppResponseDto> getParticipant(String keyword, Pageable pageable);
 
-	public Page<SearchParticipantAppResponseDto> searchParticipants(String keyword, String searchType,
-		Pageable pageable) {
-		try {
-			Page<Object> searchedPage;
-			switch (searchType.toLowerCase()) {
-				case "title":
-					searchedPage = participantRepository.searchByTitle(keyword, pageable);
-					break;
-				case "competitiontype":
-					searchedPage = participantRepository.searchByCompetitionType(keyword, pageable);
-					break;
-				case "receptiontype":
-					searchedPage = participantRepository.searchByReceptionType(keyword, pageable);
-					break;
-				case "status":
-					searchedPage = participantRepository.searchByStatus(keyword, pageable);
-					break;
-				case "uuid":
-					try {
-						UUID uuid = UUID.fromString(keyword);
-						searchedPage = participantRepository.searchByUuid(keyword, pageable);
-					} catch (IllegalArgumentException e) {
-						log.error("대회 참가자 검색 실패 로그 : {}", e.getMessage());
-						throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_SEARCH_FAIL);
-					}
-					break;
-				default:
-					throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_SEARCH_FAIL);
-			}
-			return searchedPage.map(participantMapper::toSearchParticipantAppResponseDto);
-		} catch (Exception e) {
-			log.error("검색 오류: {}, 검색 타입: {}, 키워드: {}",
-				e.getMessage(), searchType, keyword);
-			throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_SEARCH_FAIL);
-		}
-	}
+	void cancelParticipantByCompany(CancelParticipantRequestDto requestDto);
 
-	/*
-	한 명이 여러 대회 신청했을 수도 있으니 Page로 리턴
-	 */
-	public Page<SearchParticipantAppResponseDto> getParticipant(String keyword, Pageable pageable) {
-		try {
-			Page<Object> participant = participantRepository.getByUuid(UUID.fromString(keyword), pageable);
-			return participant.map(participantMapper::toSearchParticipantAppResponseDto);
+	void cancelParticipant(CancelParticipantRequestDto requestDto);
 
-		} catch (Exception e) {
-			log.error(e.getMessage());
-			throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_GET_FAIL);
-		}
-	}
-
-	@Transactional
-	public void cancelParticipantByCompany(CancelParticipantRequestDto requestDto) {
-		try {
-			CompetitionParticipantMapping participant = mappingRepository.findByCompetitionIdAndParticipantId(
-				requestDto.getCompetitionId(), requestDto.getParticipantId());
-			participant.cancel();
-		} catch (Exception e) {
-			log.error("신청 취소 디버깅 : {}", e.getMessage());
-			throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_CANCEL_FAIL);
-		}
-	}
-
-	public void cancelParticipant(CancelParticipantRequestDto requestDto) {
-		try {
-			CompetitionParticipantMapping participant = mappingRepository.findByCompetitionIdAndParticipantId(
-				requestDto.getCompetitionId(), requestDto.getParticipantId());
-			participant.cancel();
-		} catch (Exception e) {
-			log.error("신청 취소 디버깅 : {}", e.getMessage());
-			throw CustomBusinessException.from(CompetitionServiceCode.PARTICIPANT_CANCEL_FAIL);
-		}
-	}
 }
