+++ competition-service/src/test/java/com/_42195km/msa/competitionservice/CompetitionServiceImplApplicationTests.java
@@ -6,7 +6,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
 @ActiveProfiles("test")
-class CompetitionServiceApplicationTests {
+class CompetitionServiceImplApplicationTests {
 
 	@Test
 	void contextLoads() {
